olist/requirements.txt
coreapi==2.3.3
coverage==4.5.1
django==2.1.1
django-nose==1.4.5
djangorestframework==3.8.2
model-mommy==1.6.0
mysqlclient==1.3.13
protobuf==3.6.1
python-decouple==3.1
pytz==2018.5

